fix: both dev and build environments not working
