+++ .gitignore
@@ -6,3 +6,4 @@ build
 dist
 media/*
 yarn-error.log
+tsconfig.tsbuildinfo
