+++ blocks/Content/Component.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Block } from 'payload/types';
 import RichText from '../../components/RichText';
 import classes from './index.module.css';
 
@@ -8,20 +7,6 @@ export type Type = {
   blockName?: string
   content: unknown
 }
-
-export const Content: Block = {
-  slug: 'content',
-  labels: {
-    singular: 'Content',
-    plural: 'Content Blocks',
-  },
-  fields: [
-    {
-      name: 'content',
-      type: 'richText',
-    },
-  ],
-};
 
 export const Component: React.FC<Type> = (props) => {
   const { content } = props;
@@ -35,5 +20,3 @@ export const Component: React.FC<Type> = (props) => {
     </div>
   );
 };
-
-export default Content;
