+++ blocks/Content/Config.ts
@@ -1,0 +1,18 @@
+import { Block } from 'payload/types';
+
+export const Content: Block = {
+  slug: 'content',
+  labels: {
+    singular: 'Content',
+    plural: 'Content Blocks',
+  },
+  fields: [
+    {
+      name: 'content',
+      type: 'richText',
+    },
+  ],
+};
+
+
+export default Content;
