+++ blocks/Image/Component.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import NextImage from 'next/image';
-import { Block } from 'payload/types';
-import { MediaType } from '../../collections/Media';
 import RichText from '../../components/RichText';
 import classes from './index.module.css';
-import { sizes } from './sizes';
+import sizes from './sizes.json';
+import { MediaType } from '../../collections/Media';
+
 
 export type Type = {
   blockType: 'image'
@@ -13,56 +13,6 @@ export type Type = {
   caption?: any
   type: 'normal' | 'wide' | 'fullscreen'
 }
-
-export const Image: Block = {
-  slug: 'image',
-  labels: {
-    singular: 'Image',
-    plural: 'Images',
-  },
-  fields: [
-    {
-      name: 'image',
-      label: 'Image',
-      type: 'upload',
-      relationTo: 'media',
-      required: true,
-    },
-    {
-      name: 'type',
-      label: 'Type',
-      type: 'radio',
-      defaultValue: 'normal',
-      options: [
-        {
-          label: 'Normal',
-          value: 'normal',
-        },
-        {
-          label: 'Fullscreen',
-          value: 'fullscreen',
-        },
-        {
-          label: 'Wide',
-          value: 'wide',
-        },
-      ],
-      admin: {
-        layout: 'horizontal',
-      },
-    },
-    {
-      name: 'caption',
-      label: 'Caption',
-      type: 'richText',
-      admin: {
-        elements: [
-          'link',
-        ],
-      },
-    },
-  ],
-};
 
 export const Component: React.FC<Type> = (props) => {
   const { image, type, caption } = props;
