--- blocks/Image/sizes.js
@@ -1,7 +1,0 @@
-export const sizes = [
-  480,
-  1024,
-  1280,
-  1920,
-  2560,
-];
