+++ blocks/Image/sizes.json
@@ -1,0 +1,7 @@
+[
+  480,
+  1024,
+  1280,
+  1920,
+  2560
+]
