--- dev.js
@@ -1,15 +1,0 @@
-require('@babel/register')({
-  presets: [
-    'next/babel',
-  ],
-  extensions: ['.ts', '.tsx', '.js', '.jsx',
-  ],
-  env: {
-    development: {
-      sourceMaps: 'inline',
-      retainLines: true,
-    },
-  },
-});
-
-require('./server');
