+++ package.json
@@ -14,20 +14,19 @@
     "build:next": "next build",
     "build:server": "tsc --project tsconfig.server.json",
     "build:payload": "payload build",
-    "build": "cross-env NODE_ENV=production yarn build:payload && yarn build:server && cross-env NEXT_BUILD=true node dist/index.js",
-    "dev": "node dev.js",
-    "seed": "node seed/index.js",
-    "serve": "cross-env NODE_ENV=production node dist/index.js"
+    "build": "cross-env NODE_ENV=production yarn build:payload && yarn build:server && yarn serve",
+    "dev": "ts-node server.ts",
+    "seed": "ts-node seed/index.js",
+    "serve": "cross-env NODE_ENV=production PAYLOAD_CONFIG_PATH=dist/payload.config.js node dist/index.js"
   },
   "dependencies": {
-    "@babel/core": "^7.20.12",
-    "@babel/register": "^7.17.0",
     "dotenv": "^16.0.0",
     "escape-html": "^1.0.3",
     "express": "^4.18.0",
-    "next": "^12.1.0",
-    "payload": "^1.6.3",
-    "react": "^18.1.0",
+    "next": "^13.2.3",
+    "payload": "^1.6.16",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "sass": "^1.52.0",
     "slate": "^0.72.8",
     "typescript": "^4.7.0",
