+++ server.ts
@@ -6,18 +6,23 @@ import payload from 'payload';
 import { config as dotenv } from 'dotenv';
 
 dotenv({
-  path: path.resolve(__dirname, '../.env'),
+  path: path.resolve(__dirname, '.env'),
 });
 
 const dev = process.env.NODE_ENV !== 'production';
 const server = express();
 
 const start = async () => {
-  await payload.init({
-    secret: process.env.PAYLOAD_SECRET_KEY,
-    mongoURL: process.env.MONGO_URL,
-    express: server,
-  });
+  try {
+    await payload.init({
+      secret: process.env.PAYLOAD_SECRET_KEY,
+      mongoURL: process.env.MONGO_URL,
+      express: server,
+    });
+    console.log('Payload CMS started');
+  } catch (e) {
+    console.error('Error starting Payload CMS', e);
+  }
 
   if (!process.env.NEXT_BUILD) {
     const nextApp = next({ dev });
