+++ tsconfig.server.json
@@ -6,6 +6,7 @@
     "noEmit": false,
   },
   "include": [
-    "server"
+    "server.ts",
+    "payload.config.ts",
   ]
 }
